Initial commit.
Work-in-progress OSM road works tracker.  At this point it's mostly
just roundup's jinja2 template with a customised schema, minor additions
in the templates and localised text wrongly inserted directly into the
templates as a quick hack.  This is all in the name of getting a basic
thing working so I can put it out there for the users asking for it and
actually judge its usefulness before someone invests time in a proper
thing.